requirements.txt
asgiref==3.3.4; python_version >= "3.6"
certifi==2021.5.30; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
cffi==1.14.5; python_version >= "3.6"
chardet==4.0.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
cryptography==3.4.7; python_version >= "3.6"
defusedxml==0.7.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
django-allauth==0.44.0
django-cors-headers==3.7.0; python_version >= "3.6"
django-crispy-forms==1.12.0; python_version >= "3.6"
django-debug-toolbar==3.2.1; python_version >= "3.6"
django==3.2.4; python_version >= "3.6"
djangorestframework-simplejwt==4.7.1; python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
gunicorn==20.1.0; python_version >= "3.5"
idna==2.10; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
oauthlib==3.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
psycopg2-binary==2.9.1; python_version >= "3.6"
pycparser==2.20; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.4.0"
pyjwt==2.1.0; python_version >= "3.7"
python3-openid==3.2.0
pytz==2021.1; python_version >= "3.6"
requests-oauthlib==1.3.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
requests==2.25.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
sqlparse==0.4.1; python_version >= "3.6"
urllib3==1.26.5; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
whitenoise==5.2.0; python_version >= "3.5" and python_version < "4"
